refactor nodes
