+++ src/components/RichTextEditor/serializer/nodes/link.ts
@@ -1,0 +1,19 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { TLinkElement } from '@components/RichTextEditor/Plugins/LinkPlugin/types';
+import { DesignTokens } from '@components/RichTextEditor/types';
+import { TElement } from '@udecode/plate-core';
+import escapeHtml from 'escape-html';
+import { reactCssPropsToCss } from '../utils/reactCssPropsToCss';
+
+export const linkNode = (node: TElement, children: string, designTokens: DesignTokens, defaultClassNames: string) => {
+    if (node.chosenLink) {
+        const { chosenLink } = node as TLinkElement;
+        return `<a class="${defaultClassNames}" style="${reactCssPropsToCss(designTokens.link)}" target=${
+            chosenLink?.openInNewTab ? '_blank' : '_self'
+        } href="${escapeHtml(chosenLink?.searchResult?.link)}">${children}</a>`;
+    }
+    return `<a class="${defaultClassNames}" style="${reactCssPropsToCss(designTokens.link)}" href="${escapeHtml(
+        node.url as string,
+    )}">${children}</a>`;
+};
